requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
